refactor(presentation): split listRunner by two separate files
